requirements.txt
fastapi
uvicorn
sqlalchemy
dotenv
resend
pydantic[email]
argon2-cffi
asyncpg
psycopg2-binary
alembic
python-jose
httpx
python-multipart